move Lzo output file creation to LzoUtils.
