+++ src/java/com/twitter/elephantbird/mapreduce/output/LzoOutputFormat.java
@@ -2,15 +2,10 @@ package com.twitter.elephantbird.mapreduce.output;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-
-import com.hadoop.compression.lzo.LzoIndex;
 import com.hadoop.compression.lzo.LzopCodec;
+import com.twitter.elephantbird.util.LzoUtils;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FSDataOutputStream;
-import org.apache.hadoop.fs.FileStatus;
-import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
@@ -26,72 +21,14 @@ public abstract class LzoOutputFormat<K, V> extends FileOutputFormat<K, V> {
   public static final Logger LOG = LogManager.getLogger(LzoOutputFormat.class);
 
   /**
-   * A work-around to support environments with older versions of LzopCodec.
-   * It might not be feasible for to select right version of hadoop-lzo
-   * in some cases. This should be removed latest by EB-3.0.
-   */
-  private static boolean isLzopIndexSupported = false;
-  static {
-    try {
-      isLzopIndexSupported =
-        null != LzopCodec.class.getMethod("createIndexedOutputStream",
-                                          OutputStream.class,
-                                          DataOutputStream.class);
-    } catch (Exception e) {
-      // older version of hadoop-lzo.
-    }
-  }
-
-  /**
    * Helper method to create lzo output file needed to create RecordWriter
    */
   protected DataOutputStream getOutputStream(TaskAttemptContext job)
                   throws IOException, InterruptedException {
     Configuration conf = job.getConfiguration();
     LzopCodec codec = new LzopCodec();
-    codec.setConf(conf);
+    Path path = getDefaultWorkFile(job, codec.getDefaultExtension());
 
-    final Path file = getDefaultWorkFile(job, codec.getDefaultExtension());
-    final FileSystem fs = file.getFileSystem(conf);
-    FSDataOutputStream fileOut = fs.create(file, false);
-
-    FSDataOutputStream indexOut = null;
-    if (conf.getBoolean("elephantbird.lzo.output.index", false)) {
-      if ( isLzopIndexSupported ) {
-        Path indexPath = file.suffix(LzoIndex.LZO_TMP_INDEX_SUFFIX);
-        indexOut = fs.create(indexPath, false);
-      } else {
-        LOG.warn("elephantbird.lzo.output.index is enabled, but LzopCodec "
-            + "does not have createIndexedOutputStream method. "
-            + "Please upgrade hadoop-lzo.");
-      }
-    }
-
-    final boolean isIndexed = indexOut != null;
-
-    OutputStream out = ( isIndexed ?
-        codec.createIndexedOutputStream(fileOut, indexOut) :
-        codec.createOutputStream(fileOut) );
-
-
-    return new DataOutputStream(out) {
-      // override close() to handle renaming index file.
-
-      public void close() throws IOException {
-        super.close();
-
-        if ( isIndexed ) {
-          // rename or remove the index file based on file size.
-
-          Path tmpPath = file.suffix(LzoIndex.LZO_TMP_INDEX_SUFFIX);
-          FileStatus stat = fs.getFileStatus(file);
-          if (stat.getLen() <= stat.getBlockSize()) {
-            fs.delete(tmpPath, false);
-          } else {
-            fs.rename(tmpPath, file.suffix(LzoIndex.LZO_INDEX_SUFFIX));
-          }
-        }
-      }
-    };
+    return LzoUtils.getIndexedLzoOutputStream(conf, path);
   }
 }
